feat(models/construction-project): work on xml templates
For now we have two templates, the `template` which is actually used and the `createTemplateTest` which
is not actually used but kept as a reference for a working template for `addConstructionProject` for future
tinkering. We also added a ech date format helper to handlebars to format Dates to the ech required format.